MiniJVM架构重构与编译修复
主要修改：
1. 重构模块结构：将 monolithic 的 jvm.rs 拆分为多个模块化组件
2. 创建 src/lib.rs 作为库入口，导出核心模块
3. 完善错误处理体系：使用 thiserror 定义统一的 JvmError 和各子错误类型
4. 修复常量池解析：支持所有 20+ 常量类型（包括 MethodHandle、MethodType、InvokeDynamic）
5. 修复字节码解释器：实现核心指令集（算术、栈操作、控制流、方法调用）
6. 添加垃圾回收模块：实现标记-清除算法框架
7. 添加线程模块：实现线程状态管理、Monitor 同步机制
8. 创建标准库模块：为 java.lang 和 java.io 类提供 native 方法桩
9. 修复编译错误：
   - GcState/ThreadPriority 添加 Copy trait
   - InvalidConstantPoolTag 类型从 u8 改为 usize
   - 修复字符串类型不匹配问题
   - 修复 Cursor 类型和字节读取问题
   - 修复错误类型转换问题

编译状态：成功通过

+++ src/classfile/constant_pool.rs
@@ -1,5 +1,5 @@
 use std::fmt;
-use crate::error::{ClassFileError, Result};
+use crate::error::{ClassFileError, JvmError, Result};
 
 #[derive(Debug, Clone, PartialEq)]
 pub enum CpInfo {
@@ -111,7 +111,7 @@ impl ConstantPool {
                 (*class_index, *name_and_type_index),
             CpInfo::InterfaceMethodRef { class_index, name_and_type_index } => 
                 (*class_index, *name_and_type_index),
-            _ => return Err(ClassFileError::InvalidConstantPoolTag(method_ref_index as u8)),
+            _ => return Err(ClassFileError::InvalidConstantPoolTag(method_ref_index)),
         };
 
         let class_name = self.get_class_name(class_index)
@@ -123,7 +123,7 @@ impl ConstantPool {
         let (name_index, descriptor_index) = match name_and_type {
             CpInfo::NameAndType { name_index, descriptor_index } => 
                 (*name_index, *descriptor_index),
-            _ => return Err(ClassFileError::InvalidConstantPoolTag(name_and_type_index as u8)),
+            _ => return Err(ClassFileError::InvalidConstantPoolTag(name_and_type_index)),
         };
 
         let method_name = self.get_utf8(name_index)
@@ -142,7 +142,7 @@ impl ConstantPool {
         let (class_index, name_and_type_index) = match cp_info {
             CpInfo::FieldRef { class_index, name_and_type_index } => 
                 (*class_index, *name_and_type_index),
-            _ => return Err(ClassFileError::InvalidConstantPoolTag(field_ref_index as u8)),
+            _ => return Err(ClassFileError::InvalidConstantPoolTag(field_ref_index)),
         };
 
         let class_name = self.get_class_name(class_index)
@@ -154,7 +154,7 @@ impl ConstantPool {
         let (name_index, descriptor_index) = match name_and_type {
             CpInfo::NameAndType { name_index, descriptor_index } => 
                 (*name_index, *descriptor_index),
-            _ => return Err(ClassFileError::InvalidConstantPoolTag(name_and_type_index as u8)),
+            _ => return Err(ClassFileError::InvalidConstantPoolTag(name_and_type_index)),
         };
 
         let field_name = self.get_utf8(name_index)
@@ -172,7 +172,7 @@ impl ConstantPool {
         
         let utf8_index = match cp_info {
             CpInfo::String(index) => *index,
-            _ => return Err(ClassFileError::InvalidConstantPoolTag(string_index as u8)),
+            _ => return Err(ClassFileError::InvalidConstantPoolTag(string_index)),
         };
 
         self.get_utf8(utf8_index)
